move file writer into its own file
